+++ config/GALE01/splits.txt
@@ -2339,13 +2339,22 @@ melee/gm/gmregclear.c:
 	.sdata2     start:0x804DA428 end:0x804DA538
 
 melee/gm/gm_1832.c:
-	.text       start:0x8018325C end:0x8018A514
-	.rodata     start:0x803B7C28 end:0x803B7C80
-	.data       start:0x803D9248 end:0x803D9D20
-	.bss        start:0x804735A8 end:0x80473AB8
-	.sdata      start:0x804D4090 end:0x804D4158
-	.sbss       start:0x804D65F0 end:0x804D6630
-	.sdata2     start:0x804DA538 end:0x804DA650
+	.text       start:0x8018325C end:0x80188454
+	.rodata     start:0x803B7C28 end:0x803B7C68
+	.data       start:0x803D9248 end:0x803D9828
+	.bss        start:0x804735A8 end:0x80473700
+	.sdata      start:0x804D4090 end:0x804D4140
+	.sbss       start:0x804D65F0 end:0x804D6628
+	.sdata2     start:0x804DA538 end:0x804DA600
+
+melee/gm/gm_1884.c:
+	.text       start:0x80188454 end:0x8018A514
+	.rodata     start:0x803B7C68 end:0x803B7C80
+	.data       start:0x803D9828 end:0x803D9D20
+	.bss        start:0x80473700 end:0x80473AB8
+	.sdata      start:0x804D4140 end:0x804D4158
+	.sbss       start:0x804D6628 end:0x804D6630
+	.sdata2     start:0x804DA600 end:0x804DA650
 
 melee/gm/gmtoulib.c:
 	.text       start:0x8018A514 end:0x80190ABC
