+++ src/melee/gm/gm_1832.h
@@ -51,30 +51,5 @@
 /* 18838C */ void gm_8018838C_OnFrame(void);
 /* 1883C0 */ f32 gm_801883C0(void);
 /* 18841C */ bool gm_8018841C(void);
-/* 188454 */ int gm_80188454(int idx);
-/* 18846C */ int fn_8018846C(void);
-/* 1884F8 */ int fn_801884F8(void);
-/* 188550 */ void fn_80188550(int);
-/* 188644 */ void fn_80188644(void);
-/* 188738 */ void fn_80188738(HSD_JObj*);
-/* 188910 */ void fn_80188910(HSD_JObj*);
-/* 188B3C */ void fn_80188B3C(HSD_JObj*);
-/* 188D3C */ void fn_80188D3C(HSD_JObj*);
-/* 188EE8 */ void fn_80188EE8(HSD_GObj*);
-/* 1891F4 */ UNK_RET fn_801891F4(UNK_PARAMS);
-/* 189B88 */ void fn_80189B88(void);
-/* 189CDC */ void gm_80189CDC(StartMeleeData*);
-/* 18A000 */ HSD_Text* fn_8018A000(void);
-/* 18A160 */ u8 gm_8018A160(u8, u8);
-/* 18A188 */ f32 gm_8018A188(u8, u8);
-/* 18A1D8 */ f32 gm_8018A1D8(u8, u8);
-/* 18A228 */ u8 gm_8018A228(u8 difficulty, u8 stage_slot, u8 idx);
-/* 18A25C */ u8 gm_8018A25C(u8 difficulty, u8 stage_slot, u8 arg2);
-/* 18A290 */ u8 gm_8018A290(u8 difficulty, u8 stage_slot, u8 idx);
-/* 18A2C4 */ f32 gm_8018A2C4(u8, u8);
-/* 18A314 */ f32 gm_8018A314(u8, u8);
-/* 18A364 */ void fn_8018A364(int);
-/* 473814 */ extern struct TrainingMenuData gm_80473814;
-/* 473A18 */ extern UnkAllstarData gm_80473A18;
 
 #endif
