+++ src/melee/gm/gmmenu.c
@@ -3,6 +3,7 @@
 #include "gm/gmevent.h"
 
 #include <dolphin/types.h>
+#include <melee/gm/gm_1884.h>
 #include <melee/gm/gm_unsplit.h>
 #include <melee/gm/types.h>
