+++ src/melee/it/items/itheart.c
@@ -2,6 +2,7 @@
 
 #include "inlines.h"
 
+#include "gm/gm_1884.h"
 #include "gm/gm_unsplit.h"
 #include "it/inlines.h"
 #include "it/it_26B1.h"
