+++ src/melee/it/items/ittomato.c
@@ -3,7 +3,7 @@
 #include "inlines.h"
 
 #include "baselib/jobj.h"
-#include "gm/gm_1832.h"
+#include "gm/gm_1884.h"
 #include "it/inlines.h"
 #include "it/it_26B1.h"
 #include "it/it_2725.h"
